Parse 10X Genomic scRNASeq sam file (Experimental)

bc_extract_10XscSeq can extract cellular barcode, UMI and lineage barcode sequences from 10X Genomics scRNASeq bam file. This function can not process bam file directly, user needs to uncompress the bam file to get sam file in order to run this function See example.

bc_extract_10XscSeq(sam, pattern, cell_barcode_tag = "CR", umi_tag = "UR")

Arguments

sam

A string, define the un-mapped sequences

pattern

A string, define the regular expression to match the barcode sequence. The barcode sequence should be in the first catch. Please see the documents of bc_extract and example for more information.

cell_barcode_tag

A string, define the tag of 10X cell barcode field in sam file. The default is "CR".

umi_tag

A string, define the tag of UMI field in the sam file.

Value

A data.frame with 4 columns:

cell_barcode: 10X cellular barcode.

umi: UMI sequence.

barcode_seq: lineage barcode.

count: reads count.

Details

If the barcode sequence does not map to the reference genome. The user should use samtools to get the un-mapped reads and save it as sam format for using as the input. It can save a lot of time. The way to get the un-mapped reads:

samtools view -f 4 input.bam > output.sam 

Examples

## NOT run
# In the case that when the barcode sequence is not mapped to 
# reference genome, it will be much more efficiency to get 
# the un-mapped sequences as the input.

## Get un-mapped reads
# samtools view -f 4 input.bam > scRNASeq_10X.sam 

sam_file <- system.file("extdata", "scRNASeq_10X.sam", package = "CellBarcode")

bc_extract_10XscSeq(
  sam = sam_file,
  pattern = "AGATCAG(.*)TGTGGTA",
  cell_barcode_tag = "CR",
  umi_tag = "UR"
)
#>        cell_barcode          umi           barcode_seq count
#> 1  ACAACCATCACGGAGA GCTGGGATCCGG GGGAGTGGTGGAAGGTTAGGA     1
#> 2  AATCGACTCATCGCTC CTTTTGTATGGA CCTCGGCAGAGTACTTTTTAA     1
#> 3  AAGTTCGAGAGGTATT CCACGGGCACCG CGTTAATCGGTCTAGTGATTA     1
#> 4  AAGTCGTAGCACGGAT AAAGCCTGCAAG CTATATATTCGTTGTATTATA     1
#> 5  AAGACAAAGGCGCTTC CGTCTTGGTAGC TCATTTGATACTAAGAATGAA     2
#> 6  AACTTCTGTGTTCCTC GACAACCTTGGT GGGTGGGATTCAGTGATCATA     1
#> 7  AACCTTTGTCATTCCC TAATAATTGCGT CTTAACATATGTAATATCTCA     1
#> 8  AACGAAATCTGGGTCG CCAGAAACGTAC GCGGCTTCATTCTAGTGCCAA     1
#> 9  AAAGGTAAGGCCCACT AGAATTTTGTCA         ATGGTTGTAGATG     1
#> 10 AAGCAGTGGTATCAAC GCAGAGTACATG CTATATATTCGTTGTATTATA     1
#> 11 AAAGGATAGTGTGTTC ATCGTTCAAGTA CGTTAATCGGTCTAGTGATTA     1
#> 12 AAAGGGCAGCCTGAGA GAAAGTTGAGTA CGTCGATTGTGCAAACGTAGA     1
#> 13 TTCACCGATTTTGTAA TCCAGAGGTTGA TTTATCGCTTGATTTTGTATA     2
#> 14 AGGCCTCACATTCTTC TCTACCATACAT AGTGCTTTTGCTGGGTGGTGA     1
#> 15 CACAAATTTTGTAATC CAGAGGTTGATT GCGGGATCACTCGCAGGGTCA     1
#> 16 TTCGCTGATTTTGTAA TCCAGAGGTTGA AGACAACTTGGATTTTGTGTA     2
#> 17 AAACGCTAGCTGACCC CACGCTTATGTT TACCTAAGGTCGAATTTGCTA     1
#> 18 TAAGCCAAAAGAACAA AGCTGGAGGCAT         ATGGTTGTAGATG     1
#> 19 AGCAGTGGTATCAACG CAGAGTACATGG ATTGGACCATGTAATGAGTAA     1
#> 20 AAACGAAAGTTCATGC TGGTTCTGTTTC ACACATTGTTATACGTAATAA     1
#> 21 AAGGAGCTGACAGGTG GTGGCAATGCCC           TACTAGCGTAA     1
#> 22 TGAAAGTGACAACTGA GTACAGACGATG ATAGGGCATCAGAATAGTTTA     1
#> 23 AGTCTTTCGTCAAACA CGCACGACAATT         ATGGTTGTAGATG     1
#> 24 TTCACCGATTTTGTAA TCCAGAGGTTGA TCATGTTTCTGCTTCATTCCA     2
#> 25 AACACACCAATCCAGT CGGAGAGATTAC CCTACTTTAATTGAGTCAATA     1
#> 26 GGAAGCAATAGCATGA TACAAAGGCATT TAACAAACTTGACGGCCTTAA     1
#> 27 ACAAAGAGTAGGTCAG ATCTTCAAATAT                    CC     1
#> 28 AAGTGAAAGTCCCGGT AGCCATTGTCCT TTTAATAGTTTCTATTTTGGA     1
#> 29 GCAGTGGTATCAACGC AGAGTACATGGG TTTATCGCTTGATTTTGTATA     1
#> 30 AAGCAGTGGTATCAAC GCAGAGTACATG ATTGGACCATGTAATGAGTAA     2
#> 31 TAAGCCATAAACATAT TTAATAATATAT       CATGGCAAAACCAGC     1
#> 32 AAGCAGTGGTATCAAC GCAGAGTACATG TGTTCACTAGTCATAGCGACA     2
#> 33 CAAATTTTGTAATCCA GAGGTTGATTGT CTAAAGATAGTATGACTAGTA     1
#> 34 AAGCAGTGGTATCAAC GCAGAGTACATG           TAGGGTTTCAA     1
#> 35 AAGCAGTGGTATCAAC GCAGAGTACATG ATTGGTTAGAGTTCTTGATCA     1
#> 36 AGGCATTAAAGCAGCG TATCCACATAGC  GCGTTGTTGAAATTACGTAA     1
#> 37 AATAGAGGTTCTAACG AAGACAGGCTTA AGCCATTATGTCGTGTCATGA     1
#> 38 GTGGTATCAACGCAGA GTACATGGGAGC TGCTTTATGTATGTAGTGGAA     1
#> 39 TTCTTCACAGAATTGG AAAAAACTACTT         ATGGTTGTAGATG     1
#> 40 AAGACAAAGCTAAATG GTATTCTCTGAG         ATGGTTGTAGATG     1
#> 41 AAGCAGTGGTATCAAC GCAGAGTACATG AGCTTTATCCTAGTTGTCGCA     1
#> 42 AACACACTCACCGACG CTGCTCTAGGAG CGCGATCGTAGGTCCGAGTCA     1
#> 43 AAGCAGTGGTATCAAC GCAGAGTACATG GGGAGTGGTGGAAGGTTAGGA     1
#> 44 ACATGGGGTAAAAGGA GCAACATAGTTA CTGGTTTCTCGTTCCAATTAA     1
#> 45 GCAGTGGTATCAACGC AGAGTACATGGG TACCAGAGCACCTTATTTGTA     2
#> 46 CAGTGGTATCAACGCA GAGTACATGGGG TTTTGGATATTGTCCCCTAAA     1
#> 47 AAGCAGTGGTATCAAC GCAGAGTACATG  TGTGCTTAGCAGAGGGATGA     1
#> 48 GTTAAGAATACCAGTC AATCTTTCACAA TGCGTTTAAAGGCGGCTGGGA     1
#> 49 CACAACTCCTCATAAA GAGACAGCAACC CCTTTTTATTAGATCACCACA     1
#> 50 AACCCAAGGTAATTTA TAGATTCAATGC         ATGGTTGTAGATG     1
#> 51 GTGGTATCAACGCAGA GTACATGGGGCA GACAAGGGTTGGGGGACATCA     1
#> 52 AAAGTGATCTGACGCG CATCAATGACTG TATCGTTCGTTTATGTACCCA     1
#> 53 TTGGGCGTCTTTGGGC GGCCGCGTGATC AGCCATTATGTCGTGTCATGA     1
#> 54 CTACGGGAAGCAATAG CATGATACAAAG TACCTAAGGTCGAATTTGCTA     1
#> 55 TTCCTCTCAGAATTGG AAAAAATGACTT         ATGGTTGTAGATG     1
#> 56 TTCCAAATTTTGTAAT CCAGAGGTTGAT CATGCATCCAGAATTGATATA     1
#> 57 CCTTGTGAGTGTTACC ATCTTGACCCGT CATATCGTACGGTGTTAGTTA     1
#> 58 GATACAAAGGCATTAA AGCACCGTATCC TCCTATTCTGTTGGAGTACAA     1
#> 59 AAAGTCCCACTGCGAC TCGGAACCGCGT AGACAACTTGGATTTTGTGTA     1